feat : コンポーネントを分割、また投稿日時を正確に表示できるように修正　feature #226
